:poop: bad refactor separating components
